+++ app/Http/Controllers/PagesController.php
@@ -104,24 +104,30 @@ class PagesController extends Controller
         \Validator::make($request->all(),[
             "product_name" => "required|min:1|max:30",
             "deskripsi" => "min:1|max:100",
-            "tampak_depan" => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:2048',
-            "tampak_belakang" => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:2048',
+            "tampak_depan" => "dimensions:min_width=280,min_height=280,max_width=290,max_height=360",
+            "tampak_belakang" => "dimensions:min_width=280,min_height=280,max_width=290,max_height=360",
             "price" => "min:1|max:10",
             ])->validate();
 
         $product = Product::findOrFail($id);
         $product->product_name = ucfirst($request->get('product_name'));
         $product->deskripsi = ucfirst($request->get('deskripsi'));
-        if($product->tampak_depan && file_exists(storage_path('app/public/' . $product->tampak_depan))){
-            \Storage::delete('public/'.$product->tampak_depan);
-            $file = $request->file('tampak_depan')->store('tampak_depan', 'public');
-            $product->tampak_depan = $file;
-           }
-        if($product->tampak_belakang && file_exists(storage_path('app/public/' . $product->tampak_belakang))){
-            \Storage::delete('public/'.$product->tampak_belakang);
-            $file = $request->file('tampak_belakang')->store('tampak_belakang', 'public');
-            $product->tampak_belakang = $file;
-           }
+        if ($request->file('tampak_depan')){
+            if($product->tampak_depan && file_exists(storage_path('app/public/' . $product->tampak_depan))){
+                \Storage::delete('public/'.$product->tampak_depan);
+                $file = $request->file('tampak_depan')->store('tampak_depan', 'public');
+                $product->tampak_depan = $file;
+               }
+        }
+        
+        if ($request->file('tampak_belakang')){
+            if($product->tampak_belakang && file_exists(storage_path('app/public/' . $product->tampak_belakang))){
+                \Storage::delete('public/'.$product->tampak_belakang);
+                $file = $request->file('tampak_belakang')->store('tampak_belakang', 'public');
+                $product->tampak_belakang = $file;
+               }
+        }
+        
 
         $product->price = $request->get('price');
         $product->save();
